Fini ma partie en théorie

+++ app/src/main/java/com/example/tenretni/domain/models/Coordinate.kt
@@ -3,4 +3,4 @@ package com.example.tenretni.domain.models
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Coordinate (val latitude:Float = 0.0f, val longitude:Float = 0.0f)
+data class Coordinate (val latitude:Double = 0.0, val longitude:Double = 0.0)
